requirements.txt
blinker==1.8.2
click==8.1.7
daltonize==0.2.0
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
pillow==10.4.0
setuptools==71.1.0
Werkzeug==3.0.3
